Add initial setup for the test suite
This commit adds the test suite dependency, so we can plug circle CI
to run our test suite on every pull request. It also adds `gitignore`
which will ignore the files that should be added to the git repo.

This also adds the `binstub` for `rspec-core`, so we can run specs
as simple as `bin/rspec`.