Replace git sub-modules with stable releases
Initially we are using some of the behavior from mail-gpg and the
mr_smime gems. We are using git sub modules to add those repos but
they already have stable releases in rubygems.

This commit replace those sub-modules with the stable releases, and
this also cleanup some of the gem dependencies as those are already
required by the dependent gems. This commit also removes default
specs and commented out the one we need for interceptor to make the
Circle CI happy for now :)