Rename `mail-secure` gem to `enmail`
When we started we used `mail-secure` as a gem name, but later we
thought it would be better to use a different name, as the gem was
using `Mail` as a base namespace and which might confuse developer
and also create confliction with the main `mail` gem.

This commit renames the gem and also update all files accordingly,
but this still keeps the old files for now so we can use those as
reference.