Add `#sign_message` configuration option
Our main goal is to provide full control to the developer to chose
what functionality they want to use based on their requirements. This
configuration option will allow user to specify if they want the gem
to sign their message or not, by default it is set to `true` but we
can easily override it using the following configuration.

Note: We use the `configuration` as base option but it can also be
overridden while invoking any interface.

```ruby
EnMail.configuration.sign_message = true
```