Add message signing related interfaces
We need to invoke some custom interfaces to the mail instance so
we can use those directly to the mail instance to sign, encrypt
or decrypt an email message.

Instead of adding everything to the `Mail::Message` class, this
commit adds a custom module where we will defined our customize
behavior. This also implement a very simplest version of `sing`
interface which takes an optional parameter.