Add `secret_key` configuration option
By default, we are expecting user to set one secret_key that we will
use to `sign/encrypt` any message. This is just a default, so if the
user don't explicitly specify any key when invoking `sign` interface
then we will use this one.

One important thing to note, we are expecting the key as string. For
simplicity, we are not going into any of keyring implementation, all
we need from user is to pass their key as string. If the secret key
requires a passphrase then user needs to pass it when invoking the
`sign` or any other interface that will use this key.

This is the simplest implementation for now, but there are lots of
use case scenario will follow later.

```ruby
EnMail.configure do |config|
  config.secret_key = "String content for secret key"
end
```