Add signing key to message instance
By default user is configuring a `secret_key` to sign/encrypt any
messages, but they might also want to provide different key in some
use case scenario.

This commit adds the support to provide a key with `#sign` and if
they do then we are using that key as `signing_key` and if there
is no key then we will use the default one.

This also changes the `#signable?` interface, now it checks both
the `sign_message` configuration and signing key, as with out a
valid key a message can't be signed.