Retrieve the configured adapter klass
User configures `smime_adapter` by passing a symbol or string, which
is just the simplification of an actual adapter class. This commit
adds an interface to retrieve the adapter class name using the user
configuration.

One important thing to note, the returning name is a string, we did
it so we can lazily load the adapters when necessary, so please use
`Object.const_get` before invoking any method on it.