Add a key model to store key contents
The `Key` is one of the core attributes for this gem, we will use it
in different steps in the adapter, so let's organize those in more
structured way.

This commit adds a `EnMail::Key` model, which can be initialize by
passing the key contents and instead of parsing those it will keep
those content as it is. The adapter will do the necessary work to
read or use the correct key/certificate from this instance.