Add base `rubocop` style with hound config
We are trying to maintain a consistency throughout our opensource
projects. This is the first steps to follow a standard guideline
throughout our projects.

This commit adds a base `rubocop` configuration, which is pretty
much same as https://github.com/thoughtbot/guides, and we will
start adding our custom preferences ontop of that.

This commit also fixes some of the existing rubocop violations.