Implement Open PGP sign+encrypt in a combined way
The RFC 3156 explicitly allows for signing and encrypting data in
a single OpenPGP message.

The method is bit too long, but in this case it is very difficult to
shorten it.  It actually performs just five operations, everything else
are getters…  Maybe some refactoring, e.g. introduction of some
message decorator, could make lines bit shorter, and allow for splitting
them less often.  But that isn't in scope of this commit, nor pull
request.